pyproject.toml
[project]
name = "climate-knowledge-graph"
version = "0.1.0"
description = "This package  supports Knowledge Graph building for climate"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-mistralai>=0.2.11",
    "langchain-neo4j>=0.5.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
]
packages = [
    "src/climate_knowledge_graph",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"